module needs to exist on main branch to work
