+++ ft/constants.go
