+++ ft/helpers.go
@@ -58,3 +58,16 @@ func sanitizeDir(path string) string {
 }
 
 
+func equalSlices(a, b []string) bool {
+	if len(a) != len(b) {
+		return false
+	}
+	for i, v := range a {
+		if v != b[i] {
+			return false
+		}
+	}
+	return true
+}
+
+
