+++ ft/io_test.go
@@ -1,0 +1,31 @@
+package ft
+
+import (
+	"fmt"
+	"os"
+	"path/filepath"
+	"testing"
+)
+
+
+
+func TestEnsureData(t *testing.T) {
+	// Create a temporary directory for testing
+	tmpdir, err := os.MkdirTemp("", "testdata")
+	if err != nil {
+		t.Fatalf("Failed to create temp dir: %v", err)
+	}
+	defer os.RemoveAll(tmpdir)
+
+	testFilePath := filepath.Join(tmpdir, "test.bin")
+	file := ensureData(testFilePath)
+	defer file.Close()
+
+	if _, err := os.Stat(testFilePath); os.IsNotExist(err) {
+		t.Fatalf("File was not created")
+	} else {
+        fmt.Println("ensureData: Success")
+    }
+}
+
+
